add local signing test and move in testing
