+++ Cargo.toml
@@ -130,6 +130,7 @@ movement-signer-hashicorp-vault = { path = "util/signing/providers/hashicorp-vau
 movement-signer-local = { path = "util/signing/providers/local" }
 movement-signer-loader = { path = "util/signing/util/loader" }
 movement-signing-aptos = { path = "util/signing/integrations/aptos" }
+movement-signing-eth = { path = "util/signing/integrations/eth" }
 
 ## vault
 vaultrs = { version = "0.7.3" }
