+++ util/signing/integrations/eth/Cargo.toml
@@ -22,18 +22,5 @@ alloy-consensus.workspace = true
 alloy-transport-http = { workspace = true, features = ["reqwest-rustls-tls"] }
 k256 = "0.13.4"
 
-[dev-dependencies]
-aws-sdk-kms = { workspace = true }
-anyhow = { workspace = true }
-tokio = { workspace = true }
-alloy.workspace = true
-alloy-signer-aws = { git = "https://github.com/alloy-rs/alloy.git", rev = "83343b172585fe4e040fb104b4d1421f58cbf9a2" }
-aws-config = { workspace = true }
-
-ethereum-types = "0.11"
-keccak-hash = "0.10"
-hex = "0.4"
-sha3 = "0.10.8"
-
 [lints]
 workspace = true
